Replace --source-directory option with --characters option
--source-directory isn't needed now that the data is distributed with
the script.  However it was also used as a means to process a small
amount of data; this new option is a better replacement for that usage.

The README has been updated to reflect this, some files in the test
directory have also been deleted, and some tests have been modified.